pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[project]
name = "f5xc-api-fixed"
version = "0.1.0"
description = "F5 XC API Spec Validation Framework - Reconcile live API behavior with OpenAPI specs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Robin Mordasiewicz" }
]
keywords = ["f5", "xc", "openapi", "validation", "schemathesis"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.32.0",
    "pyyaml>=6.0.2",
    "openapi-spec-validator>=0.7.2",
    "jsonschema>=4.23.0",
    "httpx>=0.28.0",
    "rich>=13.9.0",
    "schemathesis>=3.28.0",
    "hypothesis>=6.100.0",
    "deepdiff>=8.0.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests>=2.32.0",
    "types-PyYAML>=6.0.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocs-minify-plugin>=0.8.0",
]

[project.scripts]
f5xc-download = "scripts.download:main"
f5xc-validate = "scripts.validate:main"
f5xc-reconcile = "scripts.reconcile:main"
f5xc-release = "scripts.release:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["scripts"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
